requirements.txt
# Core ML framework
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# Satellite data processing
rasterio>=1.3.0
gdal>=3.5.0
geopandas>=0.12.0
xarray>=2023.1.0

# TerraTorch for Prithvi backbone
terratorch>=0.1.0

# AgriFieldNet data access
radiant-mlhub>=0.5.2

# Data augmentation and processing
albumentations>=1.3.0
opencv-python>=4.7.0
pillow>=9.0.0

# Scientific computing
numpy>=1.21.0
pandas>=1.5.0
scipy>=1.9.0
scikit-learn>=1.2.0

# Configuration and logging
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.15.0
tensorboard>=2.11.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.13.0

# Utilities
tqdm>=4.64.0
rich>=13.0.0
pyyaml>=6.0
click>=8.0.0

# Development
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0